requirements.txt
annotated-types~=0.7.0
anyio~=4.6.2.post1
av~=12.3.0
bleak~=0.22.3
certifi~=2024.8.30
charset-normalizer~=3.4.0
colorama~=0.4.6
coloredlogs~=15.0.1
ctranslate2~=4.5.0
elevenlabs~=1.13.1
enum34~=1.1.10
even_glasses~=0.1.11
faster-whisper~=1.0.3
filelock~=3.16.1
flatbuffers~=24.3.25
fsspec~=2024.10.0
h11~=0.14.0
halo~=0.0.31
httpcore~=1.0.7
httpx~=0.28.0
httpx-sse~=0.4.0
huggingface-hub~=0.26.3
humanfriendly~=10.0
idna~=3.10
Jinja2~=3.1.4
joblib~=1.4.2
jsonpatch~=1.33
jsonpointer~=3.0.0
langchain-core~=0.3.21
langgraph~=0.2.53
langgraph-checkpoint~=2.0.7
langgraph-sdk~=0.1.40
langsmith~=0.1.147
llvmlite~=0.43.0
log-symbols~=0.0.14
MarkupSafe~=3.0.2
mpmath~=1.3.0
msgpack~=1.1.0
networkx~=3.4.2
numba~=0.60.0
numpy~=1.26.4
onnxruntime~=1.20.1
openwakeword~=0.6.0
orjson~=3.10.12
packaging~=24.2
protobuf~=5.29.0
pvporcupine~=1.9.5
PyAudio~=0.2.14
pydantic~=2.10.2
pydantic_core~=2.27.1
pyobjc-core~=10.3.2
pyobjc-framework-Cocoa~=10.3.2
pyobjc-framework-CoreBluetooth~=10.3.2
pyobjc-framework-libdispatch~=10.3.2
python-dotenv~=1.0.1
PyYAML~=6.0.2
RealtimeSTT~=0.3.81
redis~=5.2.0
requests~=2.32.3
requests-toolbelt~=1.0.0
scikit-learn~=1.5.2
scipy~=1.14.1
six~=1.16.0
sniffio~=1.3.1
spinners~=0.0.24
sympy~=1.13.1
tenacity~=9.0.0
termcolor~=2.5.0
threadpoolctl~=3.5.0
tokenizers~=0.21.0
torch~=2.5.1
torchaudio~=2.5.1
tqdm~=4.67.1
typing_extensions~=4.12.2
urllib3~=2.2.3
webrtcvad-wheels~=2.0.14
websocket-client~=1.8.0
websockets~=12.0
